add array capability
Changed the C API a bit so that there is only one "getter" function.
Instead, you do a "find" first, and metadata in the result of the
find tells you the type and size of data under that key. Then you
call the generic get and cast over as necessary.

Python also can deal with array values now.

For efficiency, if you are storing a non-array value, no length
is stored in the underlying data structure. However, if the length
is 0 or greater than 1, a new 2-byte length field is inserted.
This seems a reasonable overhead tradeoff.

Why would you want to store something that data of zero bytes length?
No idea, but you can.